Handlebars: deploy just the runtime (and update it)
